Properly reject `EADDRINUSE` errors
This error is not given to the `listen()` callback, but instead emitted
as an "error" event on the Server instance, so a listener needs to be
added for that event and then removed before the Promise is
resolved/rejected.

Also added `.editorconfig` from https://editorconfig.zeit.sh and added
`yarn test` command.